app/auth_verify.py
from fastapi import APIRouter, HTTPException, Request, Depends
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from jose import jwt
import os

router = APIRouter()
security = HTTPBearer()

JWT_SECRET = os.getenv("JWT_SECRET", "dev_secret")

def decode_jwt(token: str):
    try:
        decoded = jwt.decode(token, JWT_SECRET, algorithms=["HS256"])
        return decoded
    except Exception as e:
        print("❌ Erreur décodage JWT:", str(e))
        return None

def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(security)):
    token = credentials.credentials
    payload = decode_jwt(token)
    if payload is None:
        raise HTTPException(status_code=403, detail="Token invalide ou expiré")
    return payload

@router.get("/verify-token")
def verify_token(user=Depends(get_current_user)):
    return {"email": user["email"], "message": "Token valide"}